pyproject.toml
[project]
name = "marvelrivalsapi"
version = "0.1.0"
description = ""
authors = [
    {name = "externref",email = "shiva02939@gmail.com"}
]
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.9.0"
dependencies = [
    "httpx (>=0.28.1,<0.29.0)",
    "attrs (>=25.3.0,<26.0.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.8"
isort = "^6.0.1"
python-dotenv = "^1.1.0"


[tool.poetry.group.docs.dependencies]
mkdocs = {extras = ["python"], version = "^1.6.1"}
mkdocstrings = {extras = ["python"], version = "^0.29.1"}
mkdocs-material = {extras = ["imaging"], version = "^9.6.12"}

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
